Some fixes after review

+++ ewm-service/src/main/java/ru/practicum/service/event/EventServiceImpl.java
@@ -399,8 +399,8 @@ public class EventServiceImpl implements EventService {
                     .collect(Collectors.groupingBy(ParticipationRequest::getId, Collectors.summingInt(p -> 1)));
 
             for (int i = 0; i < requestList.size(); i++) {
-                if (eventsRequests.containsKey(i)) {
-                    if ((filteredEvents.get(i).getParticipantLimit() - eventsRequests.get(i)) > 0) {
+                if (eventsRequests.containsKey((long) i)) {
+                    if ((filteredEvents.get(i).getParticipantLimit() - eventsRequests.get((long) i)) > 0) {
                         resultEvents.add(events.get(i));
                     }
                 } else {
